Flag for disabling version check
